refactor(upstreams): unify duplicated HF clients into upstreams/huggingface.py (refs P2-hf)
api/routes/models.py's repo tree+metadata fetch (_fetch_hf_repo) and
api/routes/hf.py's free-text search (_fetch_hf_search) had each grown
their own httpx client setup, HF_TOKEN header injection, and response
normalizer for the same huggingface.co API. Consolidate both into
hal0.upstreams.huggingface (fetch_repo / search_models / normalise_repo_slug),
keeping route-level caching and validation in the two callers. Behavior
and response shapes are unchanged; test transport patches now target the
shared module's httpx client.
